Moves error controller to core and adds acutator endpoints
