+++ jcart-admin/src/main/java/com/sivalabs/jcart/admin/web/controllers/UserAuthController.java
@@ -1,5 +1,17 @@
-/**
- * 
+/*
+ * Copyright 2016-2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package com.sivalabs.jcart.admin.web.controllers;
 
@@ -14,8 +26,8 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
@@ -30,6 +42,7 @@ import lombok.extern.slf4j.Slf4j;
 
 /**
  * @author Siva
+ * @author rajakolli
  *
  */
 @Slf4j
@@ -50,13 +63,13 @@ public class UserAuthController extends AbstractJCartAdminController
         return "User";
     }
 
-    @RequestMapping(value = "/forgotPwd", method = RequestMethod.GET)
+    @GetMapping(value = "/forgotPwd")
     public String forgotPwd()
     {
         return VIEWPREFIX + "forgotPwd";
     }
 
-    @RequestMapping(value = "/forgotPwd", method = RequestMethod.POST)
+    @PostMapping(value = "/forgotPwd")
     public String handleForgotPwd(HttpServletRequest request,
             RedirectAttributes redirectAttributes)
     {
@@ -79,7 +92,7 @@ public class UserAuthController extends AbstractJCartAdminController
         return "redirect:/forgotPwd";
     }
 
-    @RequestMapping(value = "/resetPwd", method = RequestMethod.GET)
+    @GetMapping(value = "/resetPwd")
     public String resetPwd(HttpServletRequest request, Model model,
             RedirectAttributes redirectAttributes)
     {
@@ -102,7 +115,7 @@ public class UserAuthController extends AbstractJCartAdminController
 
     }
 
-    @RequestMapping(value = "/resetPwd", method = RequestMethod.POST)
+    @PostMapping(value = "/resetPwd")
     public String handleResetPwd(HttpServletRequest request, Model model,
             RedirectAttributes redirectAttributes)
     {
