+++ jcart-site/src/main/java/com/sivalabs/jcart/site/config/WebConfig.java
@@ -10,19 +10,16 @@ import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerFactory;
 import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory;
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.validation.Validator;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-import org.thymeleaf.extras.springsecurity4.dialect.SpringSecurityDialect;
 
 import lombok.RequiredArgsConstructor;
 
 /**
  * @author Siva
+ * @author rajakolli
  *
  */
 @Configuration
@@ -32,16 +29,6 @@ public class WebConfig extends WebMvcConfigurerAdapter
     @Value("${server.port:8443}")
     private int serverPort;
 
-    private final MessageSource messageSource;
-
-    @Override
-    public Validator getValidator()
-    {
-        LocalValidatorFactoryBean factory = new LocalValidatorFactoryBean();
-        factory.setValidationMessageSource(messageSource);
-        return factory;
-    }
-
     @Override
     public void addViewControllers(ViewControllerRegistry registry)
     {
@@ -50,12 +37,6 @@ public class WebConfig extends WebMvcConfigurerAdapter
         registry.addViewController("/register").setViewName("register");
         registry.addRedirectViewController("/", "/home");
 
-    }
-
-    @Bean
-    public SpringSecurityDialect securityDialect()
-    {
-        return new SpringSecurityDialect();
     }
 
     @Bean
