+++ jcart-site/src/main/java/com/sivalabs/jcart/site/web/models/Cart.java
@@ -1,5 +1,17 @@
-/**
- * 
+/*
+ * Copyright 2016-2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package com.sivalabs.jcart.site.web.models;
 
@@ -18,6 +30,7 @@ import lombok.Setter;
 
 /**
  * @author Siva
+ * @author rajakolli
  *
  */
 @Getter
@@ -82,12 +95,8 @@ public class Cart implements Serializable
 
     public BigDecimal getTotalAmount()
     {
-        BigDecimal amount = new BigDecimal("0.0");
-        for (LineItem lineItem : items)
-        {
-            amount = amount.add(lineItem.getSubTotal());
-        }
-        return amount;
+        return items.stream().map(LineItem::getSubTotal).reduce(BigDecimal.ZERO,
+                BigDecimal::add);
     }
 
 }
