+++ jcart-site/src/main/resources/application-default.properties
@@ -12,3 +12,5 @@ server.ssl.key-store=classpath:jcartsitekeystore.p12
 server.ssl.key-store-password=jcartsite
 server.ssl.keyStoreType=PKCS12
 server.ssl.keyAlias=jcartsitetomcat
+
+endpoints.hypermedia.enabled=true
