+++ jcart-site/src/main/resources/application-dev.properties
@@ -1,3 +1,11 @@
 
 spring.datasource.initialize=true
 spring.jpa.hibernate.ddl-auto=update
+
+management.security.enabled=false
+
+################# Info EndPoint ###################
+info.app.encoding=@project.build.sourceEncoding@
+info.app.java.source=@java.version@
+info.app.java.target=@java.version@
+management.info.git.mode=full
